AP-940 Convert to webmock
